Add hero section landing page
