+++ src/app/layout.tsx
@@ -20,8 +20,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "AImagen",
+  description: "AI Image Generator",
 };
 
 
@@ -37,12 +37,6 @@ export default function RootLayout({ children }: RootLayoutProps) {
             enableSystem
             disableTransitionOnChange
           >
-            <nav className="flex items-center justify-between">
-              <div></div>
-              <div className="m-2 p-2">
-                <ModeToggle />
-              </div>
-            </nav>
             {children}
             <Toaster/>
           </ThemeProvider>
